refactor: rename factory file from exam to answer
