+++ api/factories/__init__.py
@@ -2,6 +2,6 @@ from typing import Sequence, Type
 
 from factory.alchemy import SQLAlchemyModelFactory
 
-from api.factories.exam import AnswerFactory
+from api.factories.answer import AnswerFactory
 
 all_factories: Sequence[Type[SQLAlchemyModelFactory]] = [AnswerFactory]
